P1.3b: Split adapters handler into thin HTTP files + service helpers (#36)
internal/handlers/adapters.go was 901 LOC mixing CRUD request handling,
the /api/v1/adapters/{name}/mcp proxy hot path, Trento-specific config
parsing, and per-client MCP config building. The mix blocked Phase 2
(HTTP API -> controller-runtime shim) and made the hot path harder to
review.

Restructuring (no behavior change, no API change):

- New pkg/services/adapters/trento.go exposes ParseTrentoConfig — pure
  function, the only Trento-specific logic the handler still needed.
- New pkg/services/adapters/client_config.go owns BuildCreateClientConfig
  and BuildListClientConfig — the per-MCP-client (gemini, vscode)
  config shape builders, moved verbatim so Phase-2 can reuse them when
  the HTTP layer becomes a thin controller-runtime client shim.
- internal/handlers/adapters.go shrinks from 901 to 84 LOC: shared
  types (ErrorResponse, request/response DTOs), AdapterHandler struct
  + constructor, and the HandleAdapters dispatch shim.
- New internal/handlers/adapters_crud.go (376 LOC) holds CRUD + sync +
  health handlers. CreateAdapter now delegates Trento parsing and
  client-config building to the service helpers; ListAdapters does the
  same.
- New internal/handlers/adapters_mcp.go (314 LOC) holds the latency-
  sensitive proxy: HandleMCPProtocol, proxyToRemoteMCP, proxyToSidecar,
  rewriteSidecarURLs. Bit-identical to the previous code.

No bootstrap.go or router.go changes — splitting within package handlers
preserves every method receiver and name.

Out of scope (flagged for follow-up):
- pkg/services/adapters/service.go (1,126 LOC) is issue #8 P1.3e.
- Hardcoded http://localhost:8911 proxy base URL and the placeholder
  "Bearer adapter-session-token" — Phase-2 will derive from config /
  real token issuance.
- Awkward package services name in pkg/services/adapters/ directory.

Verification: go build ./..., go vet on changed packages, adapter
service tests pass, no new import cycles, all three handler files
under 400 LOC.

Closes #5
